refactor: renamed teacher dashboard components for consistency
